requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
colorama==0.4.3
idna==2.9
importlib-metadata==1.5.0
more-itertools==8.2.0
multidict==4.7.4
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
python-engineio==3.11.2
python-socketio==4.4.0
PyYAML==5.3
six==1.14.0
wcwidth==0.1.8
yarl==1.4.2
zipp==2.2.0
